% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transaksi.R
\docType{data}
\name{data_transaksi}
\alias{data_transaksi}
\title{Data Transaksi Pengisian Kendaraan Listrik}
\format{
Sebuah data frame dengan 1200 observasi dan 13 variabel:
\describe{
\item{id_transaksi}{ID unik transaksi}
\item{nama_stasiun}{Nama stasiun pengisian}
\item{tanggal}{Tanggal transaksi}
\item{jam_mulai}{Jam mulai pengisian}
\item{durasi_menit}{Durasi pengisian (menit)}
\item{energi_kwh}{Energi yang digunakan (kWh)}
\item{tarif_per_kwh}{Tarif listrik per kWh}
\item{status_pengisian}{Status transaksi}
\item{pendapatan}{Pendapatan transaksi}
\item{bulan_transaksi}{Nama bulan transaksi}
\item{metode_pembayaran}{Metode pembayaran}
\item{jenis_kendaraan}{Jenis kendaraan listrik}
\item{tahun_transaksi}{Tahun transaksi}
}
}
\description{
Dataset simulasi transaksi pengisian kendaraan listrik
PT Green Charge Indonesia selama bulan Juni 2026.
}
\keyword{datasets}
