% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station_kpi.R
\name{get_station_kpi}
\alias{get_station_kpi}
\title{Station KPI}
\usage{
get_station_kpi(df)
}
\arguments{
\item{df}{Data frame transaksi.}
}
\value{
Data frame KPI stasiun.
}
\description{
Menghasilkan indikator kinerja setiap stasiun pengisian.
}
