% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_tariff.R
\name{manage_tariff}
\alias{manage_tariff}
\title{Manage Charging Tariff}
\usage{
manage_tariff(df, tarif_baru)
}
\arguments{
\item{df}{Data frame transaksi.}

\item{tarif_baru}{Tarif baru per kWh.}
}
\value{
Data frame transaksi dengan tarif dan pendapatan terbaru.
}
\description{
Mengelola perubahan tarif pengisian kendaraan listrik.
}
