% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_energy.R
\name{monitor_energy_usage}
\alias{monitor_energy_usage}
\title{Monitoring Penggunaan Energi}
\usage{
monitor_energy_usage(df)
}
\arguments{
\item{df}{Data frame transaksi.}
}
\value{
Data frame ringkasan energi.
}
\description{
Membuat ringkasan penggunaan energi pada setiap stasiun pengisian.
}
