% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/receive_data.R
\name{receive_charging_data}
\alias{receive_charging_data}
\title{Receive Charging Data}
\usage{
receive_charging_data(df)
}
\arguments{
\item{df}{Data frame transaksi.}
}
\value{
Data frame yang telah divalidasi.
}
\description{
Memvalidasi data transaksi pengisian kendaraan listrik.
}
